streamlit_service/requirements.txt
absl-py==2.1.0
altair==5.4.1
astunparse==1.6.3
attrs==24.2.0
blinker==1.8.2
branca==0.7.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
et-xmlfile==1.1.0
flatbuffers==24.3.25
folium==0.17.0
fonttools==4.54.1
gast==0.6.0
geopandas==1.0.1
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.66.1
h5py==3.11.0
idna==3.10
influxdb==5.3.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.5.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
msgpack==1.1.0
namex==0.0.8
narwhals==1.8.2
numpy==1.26.4
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
padasip==1.2.2
pandas==2.2.3
patsy==0.5.6
pillow==10.4.0
protobuf==4.25.5
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
pyogrio==0.10.0
pyparsing==3.2.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.1.0
shapely==2.0.6
six==1.16.0
smmap==5.0.1
statsmodels==0.14.4
streamlit==1.39.0
streamlit-option-menu==0.3.13
streamlit_folium==0.22.1
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.4.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
watchdog==4.0.2
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
xyzservices==2024.9.0
